requirements.txt
altair==5.5.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
firebase-admin==6.6.0
frozendict==2.4.6
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
html5lib==1.1
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multitasking==0.0.11
narwhals==1.18.3
numpy==2.0.2
oauth2client==4.1.3
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
webencodings==0.5.1
yfinance==0.2.50
